Added partially working plagiarism analysis.

+++ scripts/create_corpus.py
@@ -3,6 +3,7 @@ import os
 import subprocess
 import sys
 
+from lib.plagiarism_detector.generate_ngrams import NgramsGenerator
 from scripts.create_preverticals import CreatePreverticals
 
 
@@ -10,6 +11,7 @@ class CreateCorpus:
     VERTICAL_FILES_FOLDER = CreatePreverticals.VERTICAL_FILES_FOLDER
     PRE_VERTICAL_FILES_FOLDER = CreatePreverticals.PRE_VERTICAL_FILES_FOLDER
     CONCATENATED_VERTICAL_FILE_PATH = os.path.join(VERTICAL_FILES_FOLDER, 'combined.vert')
+    NGRAM_FILES_FOLDER = 'data/analysis/ngrams/'
 
     def __init__(self):
         self.args = self.parse_commandline()
@@ -34,17 +36,22 @@ class CreateCorpus:
         for file_name in os.listdir(self.PRE_VERTICAL_FILES_FOLDER):
             pre_vertical_file_path = os.path.join(self.PRE_VERTICAL_FILES_FOLDER, file_name)
             vertical_file_path = os.path.join(self.VERTICAL_FILES_FOLDER, file_name.replace('.pvert', '.vert'))
-            if os.path.isfile(pre_vertical_file_path) and (not os.path.exists(vertical_file_path) or self.args.dont_skip):
+            if (os.path.isfile(pre_vertical_file_path)
+                    and pre_vertical_file_path.endswith('.pvert')
+                    and (not os.path.exists(vertical_file_path) or self.args.dont_skip)):
                 # Process .pvert file into .vert using "cat *.pvert | /opt/majka_pipe/majka-czech.sh > *.vert".
-                print('Processing pre-vertical %s file into %s.' % (pre_vertical_file_path, vertical_file_path), file=sys.stderr)
+                print(
+                    'Processing pre-vertical %s file into %s.' % (pre_vertical_file_path, vertical_file_path),
+                    file=sys.stderr)
 
                 # This can be easily multihreaded via:
                 # https://stackoverflow.com/questions/15107714/wait-process-until-all-subprocess-finish
                 cat_process = subprocess.Popen(['cat', pre_vertical_file_path], stdout=subprocess.PIPE)
                 with open(vertical_file_path, 'w') as vertical_file:
                     if self.args.debug:
-                        tee_process = subprocess.Popen(['tee'], stdin=cat_process.stdout, stdout=vertical_file)
-                        tee_process.wait()
+                        pass
+                        # tee_process = subprocess.Popen(['tee'], stdin=cat_process.stdout, stdout=vertical_file)
+                        # tee_process.wait()
                     else:
                         compilecorp_process = subprocess.Popen(
                             ['/opt/majka_pipe/majka-czech.sh'],
@@ -66,8 +73,31 @@ class CreateCorpus:
         #                 for line in vertical_file:
         #                     concatenated_vert_file.write(line)
 
-        print('Running compilecorp')
+        # Generate n-grams for plagiates analysis.
+        if not os.path.isdir(self.NGRAM_FILES_FOLDER):
+            os.makedirs(self.NGRAM_FILES_FOLDER)
+
+        for file_name in os.listdir(self.VERTICAL_FILES_FOLDER):
+            vertical_file_path = os.path.join(self.VERTICAL_FILES_FOLDER, file_name)
+            ngrams_file_path = os.path.join(self.NGRAM_FILES_FOLDER, file_name.replace('.vert', '.ngr'))
+            if (os.path.isfile(vertical_file_path)
+                    and vertical_file_path.endswith('.vert')
+                    and (not os.path.exists(ngrams_file_path) or self.args.dont_skip)):
+                print(
+                    'Creating ngrams %s from vertical file %s.' % (ngrams_file_path, vertical_file_path),
+                    file=sys.stderr)
+
+                with open(vertical_file_path, 'r') as vertical_file:
+                    with open(ngrams_file_path, 'w') as ngrams_file:
+                        ngrams_generator = NgramsGenerator(vertical_file, ngrams_file, 'doc', 'dbid', 's')
+                        ngrams_generator.generate()
+            else:
+                if os.path.exists(ngrams_file_path) and vertical_file_path.endswith('.vert'):
+                    print('Skipping ngram generation because of existing n-grams file %s.' % (ngrams_file_path))
+
+
         if not self.args.debug:
+            print('Running compilecorp')
             compilecorp_process = subprocess.Popen(
                 ['compilecorp', '/home/xpancik2/masters-thesis/files/compilecorp_config/dezinfo'], env=env_with_py2)
             compilecorp_process.wait()
